1. Changed prefix type in Hash and Merge traits to u64. 2. MerkleMountainRange::new() now calculate num_leaves. Also changed num_leaves from Option<usize> to usize. 3. Fixed MerkleMountainRange::find(). 4. Created MerkleMountainRange::get() and MerkleMountainRange::get_leaf() to get elements from the tree.

+++ src/mmr/partial.rs
@@ -279,9 +279,9 @@ impl<H: Merge + PartialEq + Clone, S: Store<H>> PartialMerkleMountainRange<H, S>
 
             let parent_pos = position.parent();
             let parent_hash = if position.right_node {
-                sibling_hash.merge(&hash, parent_pos.num_leaves())
+                sibling_hash.merge(&hash, parent_pos.num_leaves() as u64)
             } else {
-                hash.merge(&sibling_hash, parent_pos.num_leaves())
+                hash.merge(&sibling_hash, parent_pos.num_leaves() as u64)
             };
 
             // Add the parent to the queue.
